complete the hours/open replacement!
