+++ test/test_hours.py
@@ -1,7 +1,7 @@
 import unittest
 import sys
 sys.path.append('src')
-from hours_new import Hours  # noqa: E402
+from hours import Hours  # noqa: E402
 from discord.ext import commands  # noqa: E402
 
 client = commands.Bot(command_prefix='.')
